+++ backend/src/main/java/com/stampcrush/backend/auth/api/response/AuthTokensResponse.java
@@ -1,4 +1,4 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.api.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,14 +9,14 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthTokens {
+public class AuthTokensResponse {
 
     private String accessToken;
     private String refreshToken;
     private String grantType;
     private Long expiresIn;
 
-    public static AuthTokens of(String accessToken, String refreshToken, String grantType, Long expiresIn) {
-        return new AuthTokens(accessToken, refreshToken, grantType, expiresIn);
+    public static AuthTokensResponse of(String accessToken, String refreshToken, String grantType, Long expiresIn) {
+        return new AuthTokensResponse(accessToken, refreshToken, grantType, expiresIn);
     }
 }
