+++ backend/src/main/java/com/stampcrush/backend/auth/application/util/AuthTokensGenerator.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.application.util;
 
+import com.stampcrush.backend.auth.api.response.AuthTokensResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -8,13 +9,14 @@ import java.util.Date;
 @Component
 @RequiredArgsConstructor
 public class AuthTokensGenerator {
+
     private static final String BEARER_TYPE = "Bearer";
     private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 30;            // 30분
     private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 60 * 24 * 7;  // 7일
 
     private final JwtTokenProvider jwtTokenProvider;
 
-    public AuthTokens generate(Long memberId) {
+    public AuthTokensResponse generate(Long memberId) {
         long now = (new Date()).getTime();
         Date accessTokenExpiredAt = new Date(now + ACCESS_TOKEN_EXPIRE_TIME);
         Date refreshTokenExpiredAt = new Date(now + REFRESH_TOKEN_EXPIRE_TIME);
@@ -25,8 +27,7 @@ public class AuthTokensGenerator {
 
         String refreshToken = jwtTokenProvider.generate(subject, refreshTokenExpiredAt);
 
-
-        return AuthTokens.of(accessToken, refreshToken, BEARER_TYPE, ACCESS_TOKEN_EXPIRE_TIME / 1000L);
+        return AuthTokensResponse.of(accessToken, refreshToken, BEARER_TYPE, ACCESS_TOKEN_EXPIRE_TIME / 1000L);
     }
 
     public Long extractMemberId(String accessToken) {
