+++ backend/src/main/java/com/stampcrush/backend/auth/application/util/OAuthLoginParams.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.application.util;
 
+import com.stampcrush.backend.auth.OAuthProvider;
 import org.springframework.util.MultiValueMap;
 
 public interface OAuthLoginParams {
