+++ backend/src/main/java/com/stampcrush/backend/auth/client/KakaoInfoResponse.java
@@ -1,7 +1,8 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.client;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.stampcrush.backend.auth.OAuthProvider;
 import lombok.Getter;
 
 @Getter
