+++ backend/src/main/java/com/stampcrush/backend/auth/client/OAuthApiClient.java
@@ -1,4 +1,7 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.client;
+
+import com.stampcrush.backend.auth.OAuthProvider;
+import com.stampcrush.backend.auth.application.util.OAuthLoginParams;
 
 public interface OAuthApiClient {
