+++ backend/src/main/java/com/stampcrush/backend/auth/client/OAuthInfoResponse.java
@@ -1,9 +1,12 @@
-package com.stampcrush.backend.auth;
+package com.stampcrush.backend.auth.client;
+
+import com.stampcrush.backend.auth.OAuthProvider;
 
 public interface OAuthInfoResponse {
 
     String getNickname();
 
     String getEmail();
+
     OAuthProvider getOAuthProvider();
 }
